+++ src/presentation/tui/mod.rs
@@ -4,3 +4,4 @@ pub mod utils;
 pub mod home;
 pub mod project;
 pub mod app_runner;
+pub mod action;